requirements.txt
python-telegram-bot==20.3
telethon==1.34.0  # Critical version for legacy imports
motor==3.1.1
Flask
python-dotenv==1.0.0
pymongo==4.5.0
humanize==4.8.0